# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "description": "API для управления списком задач",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получение списка задач",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка по полям (due_date, create_date)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID задачи",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Название задачи",
                                            "type": "string",
                                            "maxLength": 255
                                        },
                                        "description": {
                                            "description": "Описание задачи",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "due_date": {
                                            "description": "Срок выполнения",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "create_date": {
                                            "description": "Дата создания",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "status": {
                                            "description": "Статус задачи",
                                            "type": "string",
                                            "enum": [
                                                "выполнена",
                                                "не выполнена"
                                            ]
                                        },
                                        "priority": {
                                            "description": "Приоритет задачи",
                                            "type": "string",
                                            "enum": [
                                                "низкий",
                                                "средний",
                                                "высокий"
                                            ]
                                        },
                                        "category": {
                                            "description": "Категория задачи",
                                            "type": "string",
                                            "maxLength": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создание задачи",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Название задачи",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "Описание задачи",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "description": "Срок выполнения",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "create_date": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "status": {
                                        "description": "Статус задачи",
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ]
                                    },
                                    "priority": {
                                        "description": "Приоритет задачи",
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ]
                                    },
                                    "category": {
                                        "description": "Категория задачи",
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получение задачи по ID",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные задачи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID задачи",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Название задачи",
                                            "type": "string",
                                            "maxLength": 255
                                        },
                                        "description": {
                                            "description": "Описание задачи",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "due_date": {
                                            "description": "Срок выполнения",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "create_date": {
                                            "description": "Дата создания",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "status": {
                                            "description": "Статус задачи",
                                            "type": "string",
                                            "enum": [
                                                "выполнена",
                                                "не выполнена"
                                            ]
                                        },
                                        "priority": {
                                            "description": "Приоритет задачи",
                                            "type": "string",
                                            "enum": [
                                                "низкий",
                                                "средний",
                                                "высокий"
                                            ]
                                        },
                                        "category": {
                                            "description": "Категория задачи",
                                            "type": "string",
                                            "maxLength": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновление задачи",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Название задачи",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "description": "Описание задачи",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "description": "Срок выполнения",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "create_date": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "status": {
                                        "description": "Статус задачи",
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ]
                                    },
                                    "priority": {
                                        "description": "Приоритет задачи",
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ]
                                    },
                                    "category": {
                                        "description": "Категория задачи",
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удаление задачи",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "API для управления задачами"
        }
    ]
}